test: remove obsolete tests and update ci workflow configurations
