+++ .github/workflows/core-tests.yml
@@ -58,3 +58,6 @@ jobs:
           bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_ownership_priority
           bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_fg_bom_field_configuration
           bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_repair_and_fg_bom
+          bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_filters
+          bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_loss_row_ownership
+          bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_supplier_allowed_items
